requirements.txt
bs4==0.0.1
Cython==0.29
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.6
gunicorn==19.9.0
llvmlite==0.25.0
lxml==4.2.5
matplotlib==3.0.1
notebook==5.7.0
numba==0.40.1
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
protobuf==3.6.1
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
Scrapy==1.5.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
tensorflow==1.9.0
